feat: add gpt-vision support
